settings: Extract config file functions to a module.
