+++ docs/contributing/code-style.md
@@ -31,7 +31,7 @@ The Vagrant setup process runs this for you.
 ## Secrets
 
 Please don't put any passwords, secret access keys, etc. inline in the
-code. Instead, use the `get_secret` function in `zproject/settings.py`
+code. Instead, use the `get_secret` function in `zproject/config.py`
 to read secrets from `/etc/zulip/secrets.conf`.
 
 ## Dangerous constructs
